requirements.txt
google-genai
playwright
supabase